feat: Add homepage route and view
The code changes include adding a new route and corresponding view for the homepage. This allows users to access the home page of the Laravel application.

Based on the recent user commits and repository commits, it seems that the focus of this commit is on adding functionality related to the homepage.